+++ pipeline/requirements.txt
@@ -62,5 +62,5 @@ mypy==1.7.1
 pre-commit==3.6.0
 
 # Security
-cryptography==41.0.8
+cryptography==45.0.5
 python-jose[cryptography]==3.3.0
