+++ pyproject.toml
@@ -1,0 +1,16 @@
+[tool.pytest.ini_options]
+testpaths = [
+    "pipeline",
+    "services", 
+    "web/src",
+    "tests"
+]
+python_files = "test_*.py"
+python_classes = "Test*"
+python_functions = "test_*"
+addopts = "--strict-markers --strict-config -v"
+markers = [
+    "asyncio: marks tests as async",
+    "integration: marks tests as integration tests",
+    "slow: marks tests as slow running"
+]
