+++ pytest.ini
@@ -1,0 +1,29 @@
+[pytest]
+# Test discovery paths - now includes adjacent tests
+testpaths = pipeline services tests
+
+# Python file patterns for test discovery
+python_files = test_*.py *_test.py
+python_classes = Test*
+python_functions = test_*
+
+# Pytest options
+addopts = --strict-markers --strict-config --verbose -ra --tb=short --maxfail=3
+
+# Async support
+asyncio_mode = auto
+
+# Test markers
+markers =
+    slow: marks tests as slow (deselect with '-m "not slow"')
+    integration: marks tests as integration tests
+    unit: marks tests as unit tests
+    e2e: marks tests as end-to-end tests
+
+# Minimum version
+minversion = 6.0
+
+# Filter warnings
+filterwarnings =
+    ignore::DeprecationWarning
+    ignore::PendingDeprecationWarning
