+++ tests/conftest.py
@@ -1,0 +1,12 @@
+"""Integration test configuration for the pipeline."""
+
+import pytest
+import asyncio
+
+
+@pytest.fixture(scope="session")
+def event_loop():
+    """Create an instance of the default event loop for the test session."""
+    loop = asyncio.get_event_loop_policy().new_event_loop()
+    yield loop
+    loop.close()
