+++ tests/test_integration.py
@@ -1,0 +1,44 @@
+"""Integration tests for the SmarterVote pipeline."""
+
+import pytest
+import asyncio
+from unittest.mock import Mock, AsyncMock
+
+# Integration tests focus on the interaction between components
+# rather than individual service unit tests (which are now adjacent to source)
+
+
+@pytest.mark.asyncio
+async def test_pipeline_integration_flow():
+    """Test the full pipeline integration flow."""
+    # This would test the actual pipeline.__main__ workflow
+    # For now, this is a placeholder for end-to-end testing
+    
+    # TODO: Implement full pipeline integration test
+    # This should test the actual 7-step workflow:
+    # DISCOVER -> FETCH -> EXTRACT -> CORPUS -> SUMMARIZE -> ARBITRATE -> PUBLISH
+    
+    assert True  # Placeholder
+
+
+@pytest.mark.asyncio 
+async def test_service_communication():
+    """Test communication between different pipeline services."""
+    # This would test how services pass data between each other
+    # and validate the data transformation at each step
+    
+    # TODO: Test actual service-to-service communication
+    # Validate schema compliance between steps
+    
+    assert True  # Placeholder
+
+
+def test_schema_validation():
+    """Test that all pipeline outputs conform to RaceJSON v0.2 schema."""
+    # This would validate that the final output matches the expected format
+    # and all required fields are present with correct types
+    
+    # TODO: Load sample race data and validate against schema
+    # Test confidence scoring, canonical issues, source attribution
+    
+    assert True  # Placeholder
